Refactor module-structure of the crate
